перенесли миграцию из плагина телеграм бота в телеграмчатлист
