+++ ThinkingHome.Plugins.TelegramBot/TelegramBotPlugin.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Telegram.Bot;
@@ -13,14 +12,9 @@ using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using ThinkingHome.Core.Plugins;
 using ThinkingHome.Core.Plugins.Utils;
-using ThinkingHome.Plugins.Database;
-
-using Chat = ThinkingHome.Plugins.TelegramBot.Model.Chat;
 
 namespace ThinkingHome.Plugins.TelegramBot {
-    public class TelegramBotPlugin(
-        DatabasePlugin database
-        ) : PluginBase, IUpdateHandler {
+    public class TelegramBotPlugin : PluginBase, IUpdateHandler {
         public event Action<Message> OnMessageReceived; 
         
         private HashSet<string> logins;
@@ -55,12 +49,6 @@ namespace ThinkingHome.Plugins.TelegramBot {
                 "register telegram message handler: {Command}", command));
         }
         
-        [DbModelBuilder]
-        public void InitModel(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<Chat>(cfg => cfg.ToTable("TelegramBot_Chat"));
-        }
-
         public override void StartPlugin()
         {
             bot.StartReceiving(this, receiverOptions, cts.Token);
@@ -87,10 +75,7 @@ namespace ThinkingHome.Plugins.TelegramBot {
             if (update.Message is { } msg) {
                 OnMessageReceived?.Invoke(update.Message);
                 if (msg.Chat.Type == ChatType.Private) {
-                    using var db = database.OpenSession();
-                    var chat = new Chat{ Id = Guid.NewGuid(), ChatId = msg.Chat.Id, Login = msg.Chat.Username};
-                    db.Set<Chat>().Add(chat);
-                    db.SaveChanges();
+                    
                     
                     if (logins.Contains(msg.Chat.Username)) {
                         var command = ParseCommand(msg.Text);
