+++ ThinkingHome.Plugins.TelegramChatList/Model/Chat.cs
@@ -1,11 +1,10 @@
-using System;
-
-namespace ThinkingHome.Plugins.TelegramBot.Model;
+namespace ThinkingHome.Plugins.TelegramChatList.Model;
 
 public class Chat {
     public Guid Id { get; set; }
     
-    public string Login { get; set; }
+    public string? Login { get; set; }
     
     public long ChatId { get; set; }
+    
 }
